# src/main/resources/META-INF/openapi.yaml
...
openapi: 3.0.3
info:
  title: Points-of-Interest Proxy / Facade API
  description: API surface exposed to interact with Backends (summit POIs + parks POIs) via the Proxy / Facade
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  #---------------------------------------------------------------------
  # TODO:
  #   Add the secure https route URL pointing to your proxy service into the setting below.
  #   You can find the route URL in RHDH by inspecting the component's details in the topology view.
  #   If you used the suggested namespace ('demo01') and component id ('poi-map') it should look like this https://demo01-poi-map-demo01.<cluster_id_here>
  #   Beware that this must be a secure HTTP URL that starts with 'https://'
  - url: https://demo-play-demo-02-demo-play.apps.demo-01-rhsys.wzhlab.top
    description: (RHDH component deployment)
  #---------------------------------------------------------------------
  - url: http://localhost:8080/
    description: (localhost testing)
paths:
...
  /backend/register/{backendId}:
    get:
      tags:
      - Backend Resource
      parameters:
      - name: backendId
        in: path
        required: true
        schema:
          type: string
      - name: endpoint
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
  /backend/unregister/{backendId}:
    get:
      tags:
      - Backend Resource
      parameters:
      - name: backendId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...